some name change and refactoring in orion
